requirements.txt
asgiref==3.5.0
attrs==21.4.0
autobahn==22.2.2
Automat==20.2.0
cffi==1.15.0
channels==3.0.4
channels-redis==3.4.*
constantly==15.1.0
cryptography==36.0.1
daphne==3.0.*
Django==4.0.*
django-crispy-forms==1.14.0
django-tinymce==3.4.*
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.0.0
pytz==2021.3
requests==2.27.*
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==22.2.0
txaio==22.2.1
typing_extensions==4.1.1
whitenoise==6.0.*
zope.interface==5.4.0
